requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
jinja2